Database Codes/requirements.txt
requests
beautifulsoup4
anthropic
python-docx
pdfplumber
schedule
